+++ src/index.ejs
@@ -5,7 +5,7 @@
   <meta name="google-site-verification" content="rT4Peb9iuBnVQdEpnAkmMx3tAWaeCUFHdHIXqLz2RFM" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <meta charset="utf-8"/>
-  <title>Tinkerware - Your Personal Devop</title>
+  <title>Tinkerware - Homepage</title>
   <link rel="shortcut icon" href="favicon.ico">
 </head>
 <body>
